Se creo un componente para mostrar el error cuando no haya conexion a internet mediante wifi o datos moviles, se implemento en los componentes que comprende a la parte de los estudiantes, students-lists, students-new, students-details, eliminar, actualizar, etc

+++ src/app/components/students/components/students-details/students-details.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { RouterExtensions } from '@nativescript/angular';
-import { Dialogs } from '@nativescript/core';
+import { Dialogs, Page } from '@nativescript/core';
 import { Student } from '@src/app/core/models/student';
 import { StudentService } from '@src/app/core/services/students/student.service';
+import { environment } from '@src/environments/environment';
 
 @Component({
   selector: 'app-students-details',
@@ -13,12 +14,15 @@ import { StudentService } from '@src/app/core/services/students/student.service'
 export class StudentsDetailsComponent implements OnInit {
   matricule: String;
   student: Student;
+  connectionActive: Boolean;
 
   constructor(
     private routerExtensions: RouterExtensions,
     private activatedRoute: ActivatedRoute,
-    private studentService: StudentService
+    private studentService: StudentService,
+    private page: Page
   ) {
+    this.verifyConnectionOnInternet();
     this.activatedRoute.params.subscribe(
       params => {
         this.matricule = params['matricule'];
@@ -37,37 +41,39 @@ export class StudentsDetailsComponent implements OnInit {
   }
 
   onDeleteTap() {
-    const confirmOptions = {
-      title: 'Eliminar estudiante', message: `¿Seguro que desea eliminar a "${this.student.name} ${this.student.lastname}"?`,
-      okButtonText: 'Sí', cancelButtonText: 'No'
-    };
-    const alertOptions = { title: 'Eliminado', message: '', okButtonText: 'Aceptar' };
-    Dialogs.confirm(confirmOptions)
-      .then(result => {
-        if (result) {
-          this.studentService.deleteStudent(this.student.matricule)
-            .subscribe(
-              (response: any) => {
-                if (response.error) {
-                  alertOptions.message = 'A ocurrido un error al eliminar el registro.';
-                  Dialogs.alert(alertOptions);
-                } else { }
-              },
-              error => {
-                console.log(error);
-              },
-              () => {
-                this.routerExtensions.navigate(['/students'], {
-                  transition: {
-                    name: 'fade'
-                  },
-                  clearHistory: true
+    if (this.verifyConnectionOnInternet()) {
+      const confirmOptions = {
+        title: 'Eliminar estudiante', message: `¿Seguro que desea eliminar a "${this.student.name} ${this.student.lastname}"?`,
+        okButtonText: 'Sí', cancelButtonText: 'No'
+      };
+      const alertOptions = { title: 'Eliminado', message: '', okButtonText: 'Aceptar' };
+      Dialogs.confirm(confirmOptions)
+        .then(result => {
+          if (result) {
+            this.studentService.deleteStudent(this.student.matricule)
+              .subscribe(
+                (response: any) => {
+                  if (response.error) {
+                    alertOptions.message = 'A ocurrido un error al eliminar el registro.';
+                    Dialogs.alert(alertOptions);
+                  } else { }
+                },
+                error => {
+                  console.log(error);
+                },
+                () => {
+                  this.routerExtensions.navigate(['/students'], {
+                    transition: {
+                      name: 'fade'
+                    },
+                    clearHistory: true
+                  });
                 });
-              });
-        }
-      }).catch(error => {
-        console.log(error);
-      });
+          }
+        }).catch(error => {
+          console.log(error);
+        });
+    }
   }
 
   onEditTap() {
@@ -76,6 +82,18 @@ export class StudentsDetailsComponent implements OnInit {
         name: 'fade'
       }
     });
+  }
+
+  verifyConnectionOnInternet(): Boolean {
+    this.connectionActive = environment.connectionActive;
+    return this.connectionActive;
+  }
+
+  tryConnection(statusConnection: any) {
+    this.connectionActive = statusConnection;
+    if (this.connectionActive) {
+      this.page.actionBarHidden = false;
+    }
   }
 
 }
